refactor: search 함수 분리
